+++ extensions/bitwarden/src/utils/hooks/useOnceEffect.ts
@@ -1,15 +1,28 @@
 import { EffectCallback, useEffect, useRef } from "react";
 
-function useOnceEffect(effect: EffectCallback, condition?: any) {
-  const ref = useRef(false);
+type AsyncEffectCallback = () => Promise<any>;
+type Effect = EffectCallback | AsyncEffectCallback;
+
+/** `useEffect` that only runs once after the `condition` is met */
+function useOnceEffect(effect: Effect, condition?: any) {
+  const hasRun = useRef(false);
 
   useEffect(() => {
-    if (ref.current) return;
+    if (hasRun.current) return;
     if (!condition) return;
 
-    ref.current = true;
+    hasRun.current = true;
+    if (isAsyncFunction(effect)) {
+      void effect();
+      return undefined;
+    }
+
     return effect();
   }, [condition]);
+}
+
+function isAsyncFunction(fn: Effect): fn is AsyncEffectCallback {
+  return fn.constructor.name === "AsyncFunction";
 }
 
 export default useOnceEffect;